Extract focus history value types
